braxton-visualizer/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "braxton-visualizer"
version = "0.1.0"
description = "Prototype tools for the Braxton Tri-Axium diagram visualizer."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy",
  "opencv-python",
  "pillow",
  "pytesseract",
  "scikit-image",
  "scipy",
  "networkx",
  "imageio",
  "tifffile",
  "lazy-loader",
  "packaging",
]
